refactor: productController.{create, update} method userId 파라미터 제거 및 UserAuthentication  파라미터 추가
